replace extra babel plugins with single plugin and extra methods
