+++ src/Framework/vite-bt-plugin/babel/blockDuplicator.ts
@@ -1,0 +1,24 @@
+import { types as t } from '@babel/core';
+
+/** Transform code block to an array of strings. */
+export function blockDuplicator(codeBlock: t.BlockStatement, sourceCode: string) {
+  // Check that function has loc info
+  if (!codeBlock.loc) {
+    throw new Error('No function body line of code property found.');
+  }
+
+  // Get line numbers of block
+  const start = codeBlock.loc.start.line;
+  const end = codeBlock.loc.end.line;
+
+  // Get code for line numbers
+  const sourceLines = sourceCode.split('\n');
+  const codeLines = sourceLines.slice(start, end - 1);
+  const codeLinesIndent = codeLines.map((line) => line.replace(/^\s{2}/, ''));
+
+  // Make the line contents array
+  const lineContentLiterals = codeLinesIndent.map((line) => t.stringLiteral(line));
+  const lineContentsArray = t.arrayExpression(lineContentLiterals);
+
+  return lineContentsArray;
+}
